Backend/data_service/requirements.txt
python-dotenv
fastapi
kafka-python
uvicorn==0.32.0
psycopg2-binary
requests
python-multipart==0.0.18
pillow
pymongo
numpy
scikit-learn